CDT-26-REFACTOR CONSULTA/MOVIMIENTOS
